Install nginx and certbot
Will auto-generate self-signed certificates for local testing. In
production (when the cookbook attribute is enabled) then certbot
will retrieve certificates for the domains also specified in the
attributes.

This has not yet been tested as the testing VM is not visible to
the internet; I will have to test this on a cloud.

A private network IP has been added for test kitchen usage to access
the VM from a browser.