Fix problem with certificates
An old root certificate expired on 2021-09-30 and some older OSes
(such as CentOS 7) are not set up to properly ignore the old root
certificate and use the replacement. This causes many downloads to
fail, even those executed by Chef Infra Client.

The solution is to install the latest `ca-certificates` package,
although I had to run the blacklist command and rebuild my test VM
as well to get downloads working again!