Revise checksum usage
SHA256 checksums must be manually generated for these files by the
Chef developer, then the checksums added to the attributes file to
match the files in the download URLs. These checksum vastly speed
up Chef Client when it is checking whether to re-download the
files.

The remote_file resource ONLY supports SHA256 checksums.
https://github.com/chef/chef/issues/3910