Add recipe and policy file for install OpenZFS
This installs the OpenZFS tools using the kABI method. This allows
the kernel and OpenZFS to be upgraded independently without
having to recompile OpenZFS each time (that happens when using the
DKMS method under Ubuntu, for example).

Once deployed, zfs and zpool should be ready to use on the node.
No pools are automatically configured.